[REFACTOR] add some best practice to adduser feature
