+++ src/app/(withNavbar)/multirole/adduser/page.tsx
@@ -6,8 +6,6 @@ import { AiOutlineArrowLeft } from "react-icons/ai";
 import { useAuth } from "@/contexts/AuthContext";
 import { InvitationEmailService, sendInvitationEmail } from "@/src/app/lib/emailservice";
 import { sanitizeInput, validateInputs } from "./utils/inputValidation";
-// import { sendInvitation } from "../adduser/services/invitationService";
-// import { InvitationPayload, InvitationResponse } from "./types/types";
 import { 
   sendInvitation, 
   InvitationPayload, 
@@ -15,7 +13,7 @@ import {
 } from "../services/invitationService";
 import { Modal } from '@/src/components/elements/modal/Modal'
 import config from "@/src/config";
-import { ConfirmModal } from "./component/confirmmodal";
+import { ConfirmModal, RoleType } from "./components/confirmmodal";
 import Dropdown from "@/src/components/Dropdown";
 
 // Role options for dropdown
@@ -214,7 +212,7 @@ export default function AddUserPage() {
         <Modal onClose={() => setShowConfirmModal(false)}>
           <ConfirmModal
             name={name}
-            role={role}
+            role={role as RoleType}
             email={email}
             onClose={() => setShowConfirmModal(false)}
             onConfirm={submitConfirmed}
